+++ backend/src/main/java/com/woowacourse/kkogkkog/coupon/application/dto/CouponSaveRequest.java
@@ -3,7 +3,7 @@ package com.woowacourse.kkogkkog.coupon.application.dto;
 import com.woowacourse.kkogkkog.coupon.domain.Coupon;
 import com.woowacourse.kkogkkog.coupon.domain.CouponStatus;
 import com.woowacourse.kkogkkog.coupon.domain.CouponType;
-import com.woowacourse.kkogkkog.domain.Member;
+import com.woowacourse.kkogkkog.member.domain.Member;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.AccessLevel;
