+++ backend/src/main/java/com/woowacourse/kkogkkog/coupon/domain/CouponEvent.java
@@ -1,6 +1,6 @@
 package com.woowacourse.kkogkkog.coupon.domain;
 
-import com.woowacourse.kkogkkog.domain.Member;
+import com.woowacourse.kkogkkog.member.domain.Member;
 import com.woowacourse.kkogkkog.exception.ForbiddenException;
 import com.woowacourse.kkogkkog.exception.InvalidRequestException;
 import java.util.function.BiConsumer;
