+++ backend/src/main/java/com/woowacourse/kkogkkog/infrastructure/event/PushAlarmEvent.java
@@ -1,9 +1,9 @@
 package com.woowacourse.kkogkkog.infrastructure.event;
 
 import com.woowacourse.kkogkkog.coupon.domain.CouponEvent;
-import com.woowacourse.kkogkkog.domain.Member;
-import com.woowacourse.kkogkkog.domain.MemberHistory;
-import com.woowacourse.kkogkkog.domain.Workspace;
+import com.woowacourse.kkogkkog.member.domain.Member;
+import com.woowacourse.kkogkkog.member.domain.MemberHistory;
+import com.woowacourse.kkogkkog.member.domain.Workspace;
 import lombok.Getter;
 
 @Getter
