+++ backend/src/main/java/com/woowacourse/kkogkkog/member/application/dto/MemberResponse.java
@@ -1,6 +1,6 @@
-package com.woowacourse.kkogkkog.application.dto;
+package com.woowacourse.kkogkkog.member.application.dto;
 
-import com.woowacourse.kkogkkog.domain.Member;
+import com.woowacourse.kkogkkog.member.domain.Member;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
