+++ backend/src/main/java/com/woowacourse/kkogkkog/member/application/dto/MemberUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.kkogkkog.application.dto;
+package com.woowacourse.kkogkkog.member.application.dto;
 
 import lombok.AccessLevel;
 import lombok.Getter;
