+++ backend/src/main/java/com/woowacourse/kkogkkog/reservation/domain/Reservation.java
@@ -3,7 +3,7 @@ package com.woowacourse.kkogkkog.reservation.domain;
 import com.woowacourse.kkogkkog.coupon.domain.Coupon;
 import com.woowacourse.kkogkkog.coupon.domain.CouponEvent;
 import com.woowacourse.kkogkkog.domain.BaseEntity;
-import com.woowacourse.kkogkkog.domain.Member;
+import com.woowacourse.kkogkkog.member.domain.Member;
 import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
