+++ backend/src/test/java/com/woowacourse/kkogkkog/common/fixture/domain/CouponFixture.java
@@ -3,7 +3,7 @@ package com.woowacourse.kkogkkog.common.fixture.domain;
 import com.woowacourse.kkogkkog.coupon.domain.Coupon;
 import com.woowacourse.kkogkkog.coupon.domain.CouponStatus;
 import com.woowacourse.kkogkkog.coupon.domain.CouponType;
-import com.woowacourse.kkogkkog.domain.Member;
+import com.woowacourse.kkogkkog.member.domain.Member;
 
 public enum CouponFixture {
