+++ backend/src/test/java/com/woowacourse/kkogkkog/common/fixture/dto/CouponDtoFixture.java
@@ -9,7 +9,7 @@ import com.woowacourse.kkogkkog.coupon.domain.CouponEvent;
 import com.woowacourse.kkogkkog.coupon.domain.CouponStatus;
 import com.woowacourse.kkogkkog.coupon.domain.CouponType;
 import com.woowacourse.kkogkkog.coupon.presentation.dto.CouponCreateRequest;
-import com.woowacourse.kkogkkog.domain.Member;
+import com.woowacourse.kkogkkog.member.domain.Member;
 import java.time.LocalDateTime;
 import java.util.List;
